+++ src/Language/GraphQL/Validate/Validation.hs
@@ -43,6 +43,7 @@ data Rule m
     | FieldRule (Field -> RuleT m)
     | ArgumentsRule (Field -> RuleT m) (Directive -> RuleT m)
     | DirectivesRule ([Directive] -> RuleT m)
+    | VariablesRule ([VariableDefinition] -> RuleT m)
 
 -- | Monad transformer used by the rules.
 type RuleT m = ReaderT (Validation m) Seq Error